Refactor phase 1 services to use aggregator API

+++ src/HslBikeApp/Models/StationHistory.cs
@@ -6,14 +6,14 @@ public record StationHistory
     public required string ArrivalStationId { get; init; }
     public string ArrivalStationName { get; init; } = "";
     public int TripCount { get; init; }
-    public int AverageDurationSeconds { get; init; }
-    public int AverageDistanceMetres { get; init; }
+    public double AverageDurationSeconds { get; init; }
+    public double AverageDistanceMetres { get; init; }
 
     public string AverageDurationFormatted
     {
         get
         {
-            var minutes = AverageDurationSeconds / 60;
+            var minutes = (int)(AverageDurationSeconds / 60);
             return minutes < 1 ? "<1 min" : $"{minutes} min";
         }
     }
